fix: attribute query handling for extension schemas
The existing serialization code supports including and excluding
attributes as required by RFC 7644 but does not handle "fully qualified"
attributes that contain the URN prefix, i.e.
"urn:ietf:params:scim:schemas:extension:enterprise:2.0:User:employeeNumber"
because it expects that `.` is used as a separator instead of `:`.

This commit adds detection of `:` and updates the existing unit tests to
capture this scenario.