fix: traversal of multi-value attribute
This patch fixes an issue with the traversal of a multi-value attribute.
If we're running in the context of a delete operation, there is the
possibility of the actual slice we're operating on to shrink in size,
possibly causing a OOB error.

This is now checked before trying to access the child property.