fix: revert enterprise extension schema change
This schema change fixed the setting of a `manager` value sent from
AzureAD, however it broke the re-sync of a user with those fields set.
AzureAD is sending a simple value for `manager`, although it is a
complex object.

My guess is that AzureAD expects the API to accept the simple value as a
reference to the `manager`, and when the user is re-fetched, the API
resolves the reference into the complex `manager` object.

This library currently does not support this, which means that all
integrations using this library cannot sync users with `manager` values
set in AzureAD.